Create base home page and clear idea project

+++ CODE/bum/src/main/java/com/smiledev/bum/repository/ProductsRepository.java
@@ -4,6 +4,8 @@ import com.smiledev.bum.entity.Products;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+
 @Repository
-public interface ProductsRepository extends JpaRepository<Products, Long> {
+public interface ProductsRepository extends JpaRepository<Products, Integer> {
+
 }
